Various cleanups and fixes
- Remove unused field from Character
- Replace include in Costume
- Reorder includes in Entity.h
- MSVC17 fix - add required return value in StateInterpolator.cpp
- replace tabs with spaces in various places.
- remove ```follower``` argument from EventFromStream
- fix InputState::serializefrom_base - m_perf_freq_diff was not set
- InputState::serializefrom now properly receives it's following
commands
- Add input command factory code in MapEventFactory
- MapInstance: fixed fallthrough in switch
- MapInstance: add command extraction/serialization for InputState
events

+++ Projects/CoX/Servers/GameServer/GameHandler.cpp
@@ -153,7 +153,7 @@ void GameHandler::on_timeout(TimerEvent *ev)
     // 2. Find all links with inactivity_time() >= disconnect_time
     //   Disconnect given link.
 
-    int timer_id = (intptr_t)ev->data();
+    intptr_t timer_id = (intptr_t)ev->data();
     switch (timer_id) {
         case Link_Idle_Timer:
             on_check_links();
